+++ project/src/main/java/lu/uni/adtool/domains/adtpredefined/MinSkill.java
@@ -86,10 +86,10 @@ public class MinSkill extends RankingDomain<BoundedInteger> implements Parametri
   /**
    * {@inheritDoc}
    *
-   * @see AdtDomain#isValueModifiable(ADTNode)
+   * @see AdtDomain#isValueModifiable(boolean)
    */
-  public final boolean isValueModifiable(ADTNode node) {
-    return node.getRole() == ADTNode.Role.PROPONENT;
+  public final boolean isValueModifiable(boolean isProponent) {
+    return isProponent;
   }
 
   /**