+++ project/src/main/java/lu/uni/adtool/domains/adtpredefined/MinTimeSeq.java
@@ -84,10 +84,10 @@ public class MinTimeSeq extends RankingDomain<RealG0> {
   /**
    * {@inheritDoc}
    *
-   * @see AdtDomain#isValueModifiable(ADTNode)
+   * @see AdtDomain#isValueModifiable(boolean)
    */
-  public final boolean isValueModifiable(ADTNode node) {
-    return node.getRole() == ADTNode.Role.PROPONENT;
+  public final boolean isValueModifiable(boolean isProponent) {
+    return isProponent;
   }
 
   /**