+++ project/src/main/java/lu/uni/adtool/domains/adtpredefined/SatOpp.java
@@ -73,8 +73,8 @@ public class SatOpp implements AdtDomain<Bool> {
    *
    * @see AdtDomain#isValueModifiable(boolean)
    */
-  public final boolean isValueModifiable(ADTNode node) {
-    return node.getRole() != ADTNode.Role.PROPONENT;
+  public final boolean isValueModifiable(boolean isProponent) {
+    return !isProponent;
   }
 
   /**