+++ project/src/main/java/lu/uni/adtool/domains/adtpredefined/SatScenario.java
@@ -75,7 +75,7 @@ public class SatScenario implements AdtDomain<Bool> {
    *
    * @see AdtDomain#isValueModifiable(boolean)
    */
-  public final boolean isValueModifiable(ADTNode node) {
+  public final boolean isValueModifiable(boolean isProponent) {
     return true;
   }