+++ project/src/main/java/lu/uni/adtool/tools/undo/EditAdtDomain.java
@@ -66,7 +66,7 @@ public class EditAdtDomain extends EditAction {
       t = domain.isOppModifiable();
       domain.setOppModifiable(this.oppModifiable);
       oppModifiable = t;
-      domainCanvas.valuesUpdated();
+      domainCanvas.valuesUpdated(false);
     }
   }